refactor(compute): rename public compute/ to private _compute/
Package renamed from braincell.compute (public) to braincell._compute
(private). Internal module files drop the leading-underscore prefix:
_point_tree.py -> topology.py, _runtime.py -> runtime.py,
_assignment_table.py -> table.py. Privacy is now expressed once at
the package boundary. No behavior change; import sites in
_multi_compartment/ and mech/_point.py docstring updated in lockstep.
