+++ braincell/_compute/topology.py
